requirements.txt
greenlet==3.1.1
playwright==1.51.0
pyee==12.1.1
python-dotenv==1.1.0
typing_extensions==4.13.2
